+++ snips_nlu/dataset/utils.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 
 from future.utils import iteritems, itervalues
+from snips_nlu_utils import normalize
 from yaml import Loader, SafeLoader
 
 from snips_nlu.constants import (
@@ -39,6 +40,27 @@ def extract_intent_entities(dataset, entity_filter=None):
                         continue
                     intent_entities[intent_name].add(chunk[ENTITY])
     return intent_entities
+
+
+def extract_entity_values(dataset, apply_normalization):
+    entities_per_intent = {intent: set() for intent in dataset[INTENTS]}
+    intent_entities = extract_intent_entities(dataset)
+    utterance_entity_values = extract_utterance_entities(dataset)
+    declared_entity_values = {
+        ent: set(ent_data[UTTERANCES])
+        for ent, ent_data in iteritems(dataset[ENTITIES])}
+    all_entity_values = {
+        ent: set(utterance_entity_values[ent]).union(
+            declared_entity_values[ent]) for ent in dataset[ENTITIES]}
+    for intent, entities in iteritems(intent_entities):
+        for entity in entities:
+            entities_per_intent[intent].update(all_entity_values[entity])
+    if apply_normalization:
+        entities_per_intent = {
+            intent: {normalize(v) for v in values}
+            for intent, values in iteritems(entities_per_intent)
+        }
+    return entities_per_intent
 
 
 def get_text_from_chunks(chunks):